int type changed to int64_t type

+++ JSONarray.hpp
@@ -15,7 +15,7 @@ namespace JsonParserVicetrice
     class JSONarray
     {
     public:
-        using VariantType = std::variant<std::string, int, bool, long double, char, std::unique_ptr<JsonParserVicetrice::JSONarray>, std::unique_ptr<JsonParserVicetrice::JSONobject>>;
+        using VariantType = std::variant<std::string, int64_t, bool, long double, char, std::unique_ptr<JsonParserVicetrice::JSONarray>, std::unique_ptr<JsonParserVicetrice::JSONobject>>;
 
         inline JSONarray() = default;
 
@@ -47,7 +47,7 @@ namespace JsonParserVicetrice
             {
                 BasicElement.push_back(*ptr);
             }
-            else if (auto ptr = std::get_if<int>(&value))
+            else if (auto ptr = std::get_if<int64_t>(&value))
             {
                 BasicElement.push_back(*ptr);
             }
